Merge asyncComputed with computed
this is a proof of concept that asyncComputed can be merged into computed and
the distinction between a normal computed and an async-computed can be made by
an extra keyword. The upside of this is that it fixes typings as we can rely on
the standard typings that come with Vue.js' computed properties.